log_analyze/requirements.txt
django==1.8.2
djangorestframework==3.1.3
django-bootstrap3==6.1.0
xlwt==1.0.0
pygments==2.0.2